Move SectorFile Related Code Into Own Folder
Move all code into dedicated folder except for code directly pertaining
to EuroScope.
